requirements.txt
numpy
pandas
seaborn
scikit-learn
mysql-connector-python
python-dotenv
scikit-learn
sqlalchemy

#-e .

